Field Names
Adding option to specify field names for the HDF5 data log in the aircraft configuration file. This works for external Mpu9250, external Bme280, SbusRx, Gps, Pitot, PressureTransducer, and Analog data. If the field name is not specified, then a default one is used. This is a big step toward making the data log completely self documenting.